agent/requirements.txt
flask
pyautogui
pillow==10.2.0
opencv-python
